clean up after multiple heredoc handled en cours
