+++ Makefile
@@ -1,14 +1,14 @@
-#******************************************************************************#
+# **************************************************************************** #
 #                                                                              #
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: smolines <smolines@student.42.fr>          +#+  +:+       +#+         #
+#    By: aubertra <aubertra@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/11/15 09:42:50 by smolines          #+#    #+#              #
-#    Updated: 2024/12/20 11:27:04 by smolines         ###   ########.fr        #
+#    Updated: 2024/12/21 14:54:37 by aubertra         ###   ########.fr        #
 #                                                                              #
-#******************************************************************************#
+# **************************************************************************** #
 
 NAME			= minishell
 
@@ -33,6 +33,7 @@ SRCS 			= 	main.c\
 					error.c\
 					exec.c\
 					exec_child.c\
+					exec_heredoc_line.c\
 					expand.c\
 					expand_cut.c\
 					expand_replace.c\
@@ -48,7 +49,8 @@ SRCS 			= 	main.c\
 					oplist_token.c\
 					parsing_dir.c\
 					parsing_heredoc.c\
-					parsing_space_operator.c\
+					parsing_operators.c\
+					parsing_operators_utils.c\
 					parsing_utils.c\
 					parsing_utils2.c\
 					parsing.c\
