+++ src/cmd_path.c
@@ -6,11 +6,11 @@
 /*   By: aubertra <aubertra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/29 15:34:19 by aubertra          #+#    #+#             */
-/*   Updated: 2024/12/09 18:34:13 by smolines         ###   ########.fr       */
+/*   Updated: 2024/12/21 14:31:32 by aubertra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//trouver le bon chemin associer avec la cmd
+/*trouver le bon chemin associer avec la cmd*/
 
 #include "minishell.h"
 #include "libft.h"
@@ -37,15 +37,15 @@ char	*find_path(char *cmd, t_env *s_env, t_manager *manager)
 	(void)manager;
 	if (!cmd || !cmd[0]) 
 		return (NULL);
-	if (!s_env || cmd[0] == '.' || cmd[0] == '/') //gerer les paths absolut/relatif et sans env
+	if (!s_env || cmd[0] == '.' || cmd[0] == '/')
 		return (absolute_path(cmd, manager));
 	paths = ft_split(get_path(s_env), ':');
 	if (!paths) 
-		return (NULL); //faire une sortie d erreur correcte si fail du split !
+		return (NULL);
 	right_path = test_path(paths, cmd, manager);
-	free_path(paths);//a ajouter dans free
+	free_path(paths);
 	if (!right_path)
-		return (NULL); //gerer la sortie d erreur si opas de right path trouver
+		return (NULL);
 	return (right_path);
 }
 
@@ -58,7 +58,7 @@ char	*join_path(char *path, char *cmd, t_manager *manager, char **paths)
 	tmp = ft_strjoin(path, "/");
 	if (!tmp)
 	{
-		free_path(paths); //pareil a ajouter au free
+		free_path(paths);
 		return (NULL);
 	}
 	to_test = ft_strjoin(tmp, cmd);
@@ -74,14 +74,13 @@ char	*join_path(char *path, char *cmd, t_manager *manager, char **paths)
 
 char	*absolute_path(char *cmd, t_manager *manager)
 {
-	(void)manager;
 	if (access(cmd, F_OK) == 0)
 	{
 		if (access(cmd, X_OK) == 0)
 			return (cmd);
 		else
 		{
-			//gestion d erreur ici
+			access_error(manager, 5, cmd);
 			return (NULL);
 		}
 	}
