+++ src/cmd_test_path.c
@@ -1,51 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   expand.c                                           :+:      :+:    :+:   */
+/*   cmd_test_path.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: aubertra <aubertra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/03 16:12:46 by smolines          #+#    #+#             */
-/*   Updated: 2024/12/13 14:18:41 by aubertra         ###   ########.fr       */
+/*   Updated: 2024/12/21 14:33:26 by aubertra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "libft.h"
 
-/*
-char	*test_path(char **paths, char *cmd, t_manager *manager)
-{
-	int		i;
-	char	*to_test;
-	char	*right_path;
+/*Test the access of the paths found*/
 
-	i = 0;
-	right_path = NULL;
-	while (paths[i])
-	{
-		to_test = join_path(paths[i], cmd, manager, paths);
-		if (to_test == NULL)
-			return (NULL);
-		if (access(to_test, F_OK) == 0)
-		{
-			if (access(to_test, X_OK) == 0)
-			{
-				if (right_path)
-					free(right_path);
-				right_path = ft_strdup(to_test);
-			}
-			else
-				right_path = NULL;
-		}
-		free(to_test);
-		i++;
-	}
-	return (right_path);
-}
-*/
-
-// test_path : créer et tester un path
+/* test_path : créer et tester un path*/
 char *create_and_check_path(char *path, char *cmd, t_manager *manager, char **paths)
 {
     char *to_test;
@@ -62,7 +32,7 @@ char *create_and_check_path(char *path, char *cmd, t_manager *manager, char **pa
     return (NULL);
 }
 
-// test_path : gère le path trouvé
+/*test_path : gère le path trouvé*/
 void update_right_path(char **right_path, char *new_path)
 {
     if (*right_path)
@@ -70,7 +40,7 @@ void update_right_path(char **right_path, char *new_path)
     *right_path = ft_strdup(new_path);
 }
 
-// tester le path
+/*tester le path*/
 char *test_path(char **paths, char *cmd, t_manager *manager)
 {
     int   i;
