+++ src/parsing.c
@@ -6,65 +6,12 @@
 /*   By: aubertra <aubertra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/28 14:46:36 by smolines          #+#    #+#             */
-/*   Updated: 2024/12/20 19:16:08 by aubertra         ###   ########.fr       */
+/*   Updated: 2024/12/21 14:55:23 by aubertra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "libft.h"
-
-/*
-//parser la chaine de caractere saisie
-int	parsing(t_manager *manager, char *line)
-{
-	int		i;
-	int		prec_space;
-
-	i = 0;
-	manager->word = NULL;
-	if (line[0] == '\0')
-		return (parsing_error(manager, 3)); //ok
-	if (only_space_symbols(line))
-		return (-1);
-	if ((count_quotes(manager, line, 34, 39) == -1) || (count_quotes(manager, line, 39, 34) % 2 == -1))
-		return (-1);
-	while (line[i])
-	{
-		manager->type = 0;
-		prec_space = 0;
-		while (line[i] && ft_is_space(line[i]))
-		{
-			prec_space++;
-			i++;
-		}
-		i = verif_operator(manager, line, i, &(manager->type));
-		// printf("type is %d i is %d, line[i] %c\n", manager->type, i, line[i]);
-		if (i == -1)
-			return (-1);
-		if (manager->type == DOUBLE_QUOTE || manager->type == SIMPLE_QUOTE)
-			i = handle_quote(line, i, manager);
-		else if (manager->type == REDIR_IN || manager->type == REDIR_OUT 
-			|| manager->type == REDIR_APPEND || manager->type == REDIR_HEREDOC)
-		{
-			if (manager->type == REDIR_APPEND || manager->type == REDIR_HEREDOC)
-				i++;
-			i = handle_redir(manager, line, i);
-		}
-		else if (manager->type == PIPE)
-			i = handle_pipe(manager, line, i);
-		else if (manager->type == DIR)
-			i = handle_dir(manager, line, i, NULL);
-		else if (manager->type == ENV_VAR)
-			i = handle_env_pars(manager, line, i);
-		else
-			i = regular_word(manager, line, i);
-		if (i == -1)
-			return (free(manager->word), -1);
-		token_add_back(&(manager->token_first), token_new(prec_space, manager));
-		free(manager->word);
-	}
-	return (0);
-}*/
 
 // parsing : Gère les cas d'erreur dans la ligne d'entrée
 int handle_parsing_errors(t_manager *manager, char *line) 
@@ -77,7 +24,6 @@ int handle_parsing_errors(t_manager *manager, char *line)
         return (-1);
     return (0);
 }
-
 
 // parsing : Traite les espaces dans la ligne
 int skip_spaces(char *line, int i, int *prec_space)
