+++ src/parsing_operators.c
@@ -1,46 +1,17 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parsing_space_operator.c                           :+:      :+:    :+:   */
+/*   parsing_operators.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: aubertra <aubertra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/28 14:46:36 by smolines          #+#    #+#             */
-/*   Updated: 2024/12/19 14:26:56 by aubertra         ###   ########.fr       */
+/*   Updated: 2024/12/21 14:53:01 by aubertra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "libft.h"
-
-int	is_symbols(char c)
-{
-	if (c == ':' || c == '!')
-		return (1);
-	return (0);
-}
-
-int	only_space_symbols(char *str)
-{
-	while (*str)
-	{
-		if (!ft_is_space(*str) && !is_symbols(*str))
-			return (0);
-		str++;	
-	}
-	return (1);
-}
-
-int	only_space(char *str)
-{
-	while (*str)
-	{
-		if (!ft_is_space(*str))
-			return (0);
-		str++;	
-	}
-	return (1);
-}
 
 //recuperer un token "operator"
 int	is_operators(t_manager *manager, char *line, int i)
